Conflicts:
	profile2/pom.xml

SAK-29398

+++ profile2/tool/src/java/org/sakaiproject/profile2/tool/pages/panels/MessageThreadsView.java
@@ -87,7 +87,7 @@ public class MessageThreadsView extends Panel {
 		
 		//get our list of messages
 		final MessageThreadsDataProvider provider = new MessageThreadsDataProvider(currentUserUuid);
-		int numMessages = provider.size();
+		int numMessages = (int) provider.size();
 		
 		//message list
 		DataView<MessageThread> messageThreadList = new DataView<MessageThread>("messageThreadList", provider) {
